requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
python-dotenv==1.0.0